+++ R/emuR-requery.database.R
@@ -46,6 +46,7 @@ drop_requeryTmpTables <- function(emuDBhandle){
 ##' @param ignoreOutOfBounds ignore result segments that are out of bundle bounds
 ##' @param calcTimes calculate times for resulting segments (results in \code{NA} values for start and end times in emuseg/emuRsegs). As it can be very computationally expensive to 
 ##' calculate the times for large nested hierarchies it can be turned of via this boolian parameter. 
+##' @param timeRefSegmentLevel set time segment level from which to derive time information. It is only necessary to set this parameter if more than one child level contains time information and the queried parent level is of type ITEM.
 ##' @return result set object of class 'emuRsegs' containing the requeried segments
 ##' @export
 ##' @seealso \code{\link{query}} \code{\link{requery_hier}} \code{\link{emuRsegs}}
@@ -93,7 +94,9 @@ drop_requeryTmpTables <- function(emuDBhandle){
 ##' requery_seq(ae, sl3, length = 3, ignoreOutOfBounds = TRUE)
 ##' 
 ##' }
-requery_seq<-function(emuDBhandle, seglist, offset=0,offsetRef='START',length=1,ignoreOutOfBounds=FALSE, calcTimes = T){
+requery_seq<-function(emuDBhandle, seglist, offset = 0, offsetRef = 'START', 
+                      length = 1, ignoreOutOfBounds = FALSE, calcTimes = T, timeRefSegmentLevel = NULL){
+  
   if(!inherits(seglist,"emuRsegs")){
     stop("Segment list 'seglist' must be of type 'emuRsegs'. (Do not set a value for 'resultType' parameter in the query() command; then the default resultType=emuRsegs will be used)")
   }
@@ -151,7 +154,7 @@ requery_seq<-function(emuDBhandle, seglist, offset=0,offsetRef='START',length=1,
     DBI::dbWriteTable(emuDBhandle$connection, "interm_res_items_tmp_root", he, overwrite=T)
     
     
-    trSl=convert_queryResultToEmuRsegs(emuDBhandle, timeRefSegmentLevel = NULL, filteredTablesSuffix = "", queryStr = "FROM REQUERY", calcTimes)
+    trSl=convert_queryResultToEmuRsegs(emuDBhandle, timeRefSegmentLevel = timeRefSegmentLevel, filteredTablesSuffix = "", queryStr = "FROM REQUERY", calcTimes)
     drop_allTmpTablesDBI(emuDBhandle)
     
     return(trSl)
@@ -171,7 +174,8 @@ requery_seq<-function(emuDBhandle, seglist, offset=0,offsetRef='START',length=1,
 ##' @param level character string: name of target level
 ##' @param collapse collapse the found items in the requested level to a sequence (concatenated with ->). If set to \code{FALSE} separate items as new entries in the emuRsegs object are returned.
 ##' @param calcTimes calculate times for resulting segments (results in \code{NA} values for start and end times in emuseg/emuRsegs). As it can be very computationally expensive to 
-##' calculate the times for large nested hierarchies it can be turned of via this boolian parameter. 
+##' calculate the times for large nested hierarchies it can be turned of via this boolian parameter.
+##' @param timeRefSegmentLevel set time segment level from which to derive time information. It is only necessary to set this parameter if more than one child level contains time information and the queried parent level is of type ITEM.
 ##' @param verbose be verbose
 ##' @return result set object of class \link{emuRsegs}
 ##' @export
@@ -211,7 +215,9 @@ requery_seq<-function(emuDBhandle, seglist, offset=0,offsetRef='START',length=1,
 ##' requery_seq(ae, requery_hier(ae, sl1, level = 'Phoneme'), offsetRef = 'END')
 ##' 
 ##' }
-requery_hier<-function(emuDBhandle, seglist, level, collapse = TRUE, calcTimes = T, verbose = TRUE){
+requery_hier<-function(emuDBhandle, seglist, level, collapse = TRUE, 
+                       calcTimes = T, timeRefSegmentLevel = NULL, verbose = TRUE){
+  
   if(!inherits(seglist,"emuRsegs")){
     stop("Segment list 'seglist' must be of type 'emuRsegs'. (Do not set a value for 'resultType' parameter for the query, the default resultType will be used)")
   }
@@ -328,7 +334,7 @@ requery_hier<-function(emuDBhandle, seglist, level, collapse = TRUE, calcTimes =
       
     }
     # trSl=convert_queryResultToVariableEmuRsegs(emuDBhandle)
-    trSl=convert_queryResultToEmuRsegs(emuDBhandle, timeRefSegmentLevel = NULL, filteredTablesSuffix = "", queryStr = "FROM REQUERY", calcTimes)
+    trSl=convert_queryResultToEmuRsegs(emuDBhandle, timeRefSegmentLevel = timeRefSegmentLevel, filteredTablesSuffix = "", queryStr = "FROM REQUERY", calcTimes)
     inSlLen=nrow(seglist)
     trSlLen=nrow(trSl)